+++ react/components/atoms/index.js
@@ -2,6 +2,5 @@ export * from "./avatar";
 export * from "./button";
 export * from "./button-keyboard";
 export * from "./button-tab";
-export * from "./key-value";
 export * from "./icon";
 export * from "./tag";
