+++ react/components/molecules/index.js
@@ -1,4 +1,5 @@
 export * from "./header";
+export * from "./key-value";
 export * from "./key-values";
 export * from "./keyboard-numeric";
 export * from "./section-view";
